+++ repo/src/main/java/com/parashift/onlyoffice/scripts/DownloadAs.java
@@ -35,7 +35,7 @@ import java.util.*;
 import java.util.zip.Deflater;
 
 /*
-    Copyright (c) Ascensio System SIA 2023. All rights reserved.
+    Copyright (c) Ascensio System SIA 2024. All rights reserved.
     http://www.onlyoffice.com
 */
 @Component(value = "webscript.onlyoffice.download-as.post")