+++ repo/src/main/java/com/parashift/onlyoffice/scripts/History.java
@@ -17,7 +17,7 @@ import java.io.*;
 import java.util.*;
 
 /*
-    Copyright (c) Ascensio System SIA 2023. All rights reserved.
+    Copyright (c) Ascensio System SIA 2024. All rights reserved.
     http://www.onlyoffice.com
 */
 @Component(value = "webscript.onlyoffice.history.get")