+++ repo/src/main/java/com/parashift/onlyoffice/sdk/manager/document/DocumentManagerImpl.java
@@ -2,6 +2,7 @@ package com.parashift.onlyoffice.sdk.manager.document;
 
 import com.onlyoffice.manager.document.DefaultDocumentManager;
 import com.onlyoffice.manager.settings.SettingsManager;
+import com.parashift.onlyoffice.util.Util;
 import org.alfresco.model.ContentModel;
 import org.alfresco.service.cmr.coci.CheckOutCheckInService;
 import org.alfresco.service.cmr.repository.NodeRef;
@@ -15,13 +16,11 @@ import java.io.Serializable;
 import java.util.Map;
 
 /*
-   Copyright (c) Ascensio System SIA 2023. All rights reserved.
+   Copyright (c) Ascensio System SIA 2024. All rights reserved.
    http://www.onlyoffice.com
 */
 
 public class DocumentManagerImpl extends DefaultDocumentManager {
-    public static final QName EditingKeyAspect = QName.createQName("onlyoffice:editing-key");
-    public static final QName EditingHashAspect = QName.createQName("onlyoffice:editing-hash");
 
     @Autowired
     @Qualifier("checkOutCheckInService")
@@ -40,7 +39,7 @@ public class DocumentManagerImpl extends DefaultDocumentManager {
 
         String key = null;
         if (cociService.isCheckedOut(nodeRef)) {
-            key = (String) nodeService.getProperty(cociService.getWorkingCopy(nodeRef), EditingKeyAspect);
+            key = (String) nodeService.getProperty(cociService.getWorkingCopy(nodeRef), Util.EditingHashAspect);
         }
 
         if (key == null) {