+++ repo/src/main/java/com/parashift/onlyoffice/util/Util.java
@@ -24,7 +24,7 @@ import java.security.SecureRandom;
 import java.util.*;
 
 /*
-   Copyright (c) Ascensio System SIA 2023. All rights reserved.
+   Copyright (c) Ascensio System SIA 2024. All rights reserved.
    http://www.onlyoffice.com
 */