+++ share/src/main/java/com/onlyoffice/web/evaluator/IsFillForm.java
@@ -1,15 +1,24 @@
 /*
-   Copyright (c) Ascensio System SIA 2023. All rights reserved.
+   Copyright (c) Ascensio System SIA 2024. All rights reserved.
    http://www.onlyoffice.com
 */
 package com.onlyoffice.web.evaluator;
 
+import com.onlyoffice.model.common.Format;
+import com.onlyoffice.web.scripts.OnlyofficeSettingsQuery;
 import org.alfresco.error.AlfrescoRuntimeException;
 import org.alfresco.web.evaluator.BaseEvaluator;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
 public class IsFillForm extends BaseEvaluator {
+
+    private OnlyofficeSettingsQuery onlyofficeSettings;
+
+    public void setOnlyofficeSettings(OnlyofficeSettingsQuery onlyofficeSettings) {
+        this.onlyofficeSettings = onlyofficeSettings;
+    }
+
     @Override
     public boolean evaluate(JSONObject jsonObject) {
         try {
@@ -17,7 +26,15 @@ public class IsFillForm extends BaseEvaluator {
             if (fileName != null) {
                 String docExt = fileName.substring(fileName.lastIndexOf(".") + 1).trim().toLowerCase();
 
-                if (docExt.equals("oform")) return true;
+                boolean isFillable = false;
+
+                for (Format format : onlyofficeSettings.getSupportedFormats()) {
+                    if (format.getName().equals(docExt) && format.getActions().contains("fill")) {
+                        isFillable = true;
+                    }
+                }
+
+                return isFillable;
             }
         } catch (Exception err) {
             throw new AlfrescoRuntimeException("Failed to run action evaluator", err);